# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API davidWebProject",
        "description": "Cette API est conçue pour gérer l'intégration de diverses fonctionnalités dans mon projet de freelance. Elle permet la gestion complète du backend, incluant la création, la mise à jour, la suppression et la récupération de données essentielles telles que les témoignages clients, les projets, les utilisateurs, et d'autres éléments de contenu. L'API alimente mon site web en fournissant les données nécessaires tout en facilitant la gestion administrative des opérations. Elle est sécurisée via l'authentification par jeton et est optimisée pour une utilisation par des applications web et mobiles. Idéale pour un usage interne, cette API centralise et simplifie la gestion de mon activité freelance.",
        "contact": {
            "name": "Support",
            "url": "https://www.davidwebprojects.fr",
            "email": "contact@davidwebprojects.fr"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://apifreelance.davidwebprojects.fr",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/civilities": {
            "get": {
                "tags": [
                    "Civilities"
                ],
                "summary": "Récupérer la liste des civilités",
                "operationId": "getCivilities",
                "responses": {
                    "200": {
                        "description": "Liste des civilités",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Civility"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Civilities"
                ],
                "summary": "Créer une nouvelle civilité",
                "operationId": "createCivility",
                "requestBody": {
                    "description": "Données nécessaires pour créer une civilité",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "shortLabel",
                                    "longLabel"
                                ],
                                "properties": {
                                    "shortLabel": {
                                        "type": "string",
                                        "example": "Mr"
                                    },
                                    "longLabel": {
                                        "type": "string",
                                        "example": "Monsieur"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Civilité créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Civility"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/civilities/{id}": {
            "get": {
                "tags": [
                    "Civilities"
                ],
                "summary": "Récupérer une civilité par ID",
                "operationId": "getCivilityById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la civilité",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la civilité",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Civility"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Civilité non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Civilities"
                ],
                "summary": "Mettre à jour une civilité",
                "operationId": "updateCivility",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la civilité à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données nécessaires pour mettre à jour une civilité",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "shortLabel": {
                                        "type": "string",
                                        "example": "Mr"
                                    },
                                    "longLabel": {
                                        "type": "string",
                                        "example": "Monsieur"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Civilité mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Civility"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Civilité non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Civilities"
                ],
                "summary": "Supprimer une civilité",
                "operationId": "deleteCivility",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la civilité à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Civilité supprimée avec succès"
                    },
                    "404": {
                        "description": "Civilité non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Civility": {
                "title": "Civilité",
                "description": "Modèle de civilité",
                "required": [
                    "shortLabel",
                    "longLabel"
                ],
                "properties": {
                    "id": {
                        "description": "ID unique de la civilité",
                        "type": "integer"
                    },
                    "shortLabel": {
                        "description": "Abréviation de la civilité",
                        "type": "string"
                    },
                    "longLabel": {
                        "description": "Nom complet de la civilité",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}